auth - fix login form blink (move user and isAuthenticated to authKeep)
