+++ src/store/store-moderator.js
@@ -5,7 +5,8 @@ export default function configureModerator(store, router) {
     switch (type) {
       case 'auth/logout':
         // return store.commit('booking/resetState');
-        store.dispatch('memo/clear_form_data')
+        store.dispatch('memo/clear_form_data');
+        store.dispatch('authKeep/clear');
     }
     /* eslint-enable */
   });
